core/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "core"
version = "0.1.0"
description = "Shared utilities"
requires-python = ">=3.12"
readme = "README.md"
authors = [{ name = "Your Name", email = "you@example.com" }]
dependencies = [
    "pydantic>=2.7",
    "structlog>=24.1.0",
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.9",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
